pyproject.toml
[project]
name = "openwhispersync"
version = "0.1.0"
description = "Lightweight audiobook alignment using whisper"
authors = [
    {name = "Ian", email = "itaylo19@students.kennesaw.edu"},
    {name = "Droan", email = "dpatel244@students.kennesaw.edu"}
]
dependencies = [
    "openai-whisper>=20231117",
    "ebooklib>=0.18.0",
    "rapidfuzz>=3.6.1",
    "click>=8.1.7",
    "rich>=13.7.0",
    "beautifulsoup4>=4.12.0",
    "pydub>=0.25.1",
    "numpy>=1.24.0",
    "librosa>=0.10.1",
    "matplotlib>=3.8.0"
]
requires-python = ">=3.9"

[project.scripts]
openwhispersync = "openwhispersync.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build" 